chore: 코멘트 반영 수정(서비스 내 코드 책임 분리) 및 안쓰는 파일 삭제
